Revert "Merge branch 'tests/2017_bd_full_tests' into release_2.1"
This reverts commit 276a7da2c3d21b6b8625f7e98fde4586568404ad, reversing
changes made to 7988fa14f45e48862f36672e2e776c42f538c83b.
