+++ src/test/java/gwt/material/design/client/base/TableBaseTest.java
@@ -63,8 +63,33 @@ public class TableBaseTest extends GWTTestCase {
         return "gwt.material.design.GwtMaterialTable";
     }
 
-    public <T extends MaterialDataTable<Person>> void test(T table) {
+    @Override
+    protected void gwtSetUp() throws Exception {
+        super.gwtSetUp();
+        setup();
+    }
+
+    public void setup() {
+        WithJQueryResources jQueryResources = GWT.create(WithJQueryResources.class);
+        // Test JQuery
+        MaterialDesign.injectJs(jQueryResources.jQuery());
+        assertTrue(MaterialDesign.isjQueryLoaded());
+        // Test Materialize
+        MaterialDesign.injectJs(MaterialResources.INSTANCE.materializeJs());
+        assertTrue(MaterialDesign.isMaterializeLoaded());
+        // Inject Resources
+        MaterialDesign.injectJs(MaterialTableBundle.INSTANCE.jQueryExt());
+        MaterialDesign.injectJs(MaterialTableBundle.INSTANCE.stickyth());
+        MaterialDesign.injectJs(MaterialTableBundle.INSTANCE.tableSubHeaders());
+        MaterialDesign.injectJs(MaterialTableBundle.INSTANCE.greedyScroll());
+        // gwt-material-jquery Test
+        assertNotNull($("body"));
+    }
+
+    public <T extends MaterialDataTable<Person>> void checkTable(T table) {
         setup(table);
+        checkEventHandlers(table);
+        checkColumnsRows(table);
         checkStructure(table);
         checkStretch(table);
         checkDynamicColumn(table);
@@ -74,10 +99,10 @@ public class TableBaseTest extends GWTTestCase {
     public <T extends MaterialDataTable<Person>> void setup(T table) {
         final String HEIGHT = "calc(100vh - 131px)";
         table.setHeight(HEIGHT);
-        table.setUseStickyHeader(true);
-        assertTrue(table.isUseStickyHeader());
-        table.setUseStickyHeader(false);
-        assertFalse(table.isUseStickyHeader());
+        //table.setUseStickyHeader(true);
+        //assertTrue(table.isUseStickyHeader());
+        //table.setUseStickyHeader(false);
+        //assertFalse(table.isUseStickyHeader());
         table.setUseCategories(true);
         assertTrue(table.isUseCategories());
         table.setUseCategories(false);
@@ -97,6 +122,237 @@ public class TableBaseTest extends GWTTestCase {
         table.setRowFactory(new PersonRowFactory());
         table.setCategoryFactory(new CustomCategoryFactory());
         table.setRenderer(new CustomRenderer<>());
+    }
+
+    public <T extends MaterialDataTable<Person>> void checkEventHandlers(T table) {
+        // Category Closed
+        final boolean[] isCategoryClosedFired = {false};
+        table.addCategoryClosedHandler((e, param1) -> {
+            isCategoryClosedFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.CATEGORY_CLOSED, null);
+        assertTrue(isCategoryClosedFired[0]);
+        // Category Opened
+        final boolean[] isCategoryOpenedFired = {false};
+        table.addCategoryOpenedHandler((e, param1) -> {
+            isCategoryOpenedFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.CATEGORY_OPENED, null);
+        assertTrue(isCategoryOpenedFired[0]);
+        // Row Context Menu
+        final boolean[] isRowContextMenuFired = {false};
+        table.addRowContextMenuHandler((e, param1, param2, param3) -> {
+            isRowContextMenuFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_CONTEXTMENU, null);
+        assertTrue(isRowContextMenuFired[0]);
+        // Row Count Change
+        final boolean[] isRowCountChangeFired = {false};
+        table.addRowCountChangeHandler((e, param1, param2) -> {
+            isRowCountChangeFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_COUNT_CHANGE, null);
+        assertTrue(isRowCountChangeFired[0]);
+        // Row Double Click
+        final boolean[] isRowDoubleClickFired = {false};
+        table.addRowDoubleClickHandler((e, param1, param2, param3) -> {
+            isRowDoubleClickFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_DOUBLECLICK, null);
+        assertTrue(isRowDoubleClickFired[0]);
+        // Row Expand
+        final boolean[] isRowExpandFired = {false};
+        table.addRowExpandHandler((e, param1) -> {
+            isRowExpandFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_EXPAND, null);
+        assertTrue(isRowExpandFired[0]);
+        // Row Expanded
+        final boolean[] isRowExpandedFired = {false};
+        table.addRowExpandedHandler((e, param1) -> {
+            isRowExpandedFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_EXPANDED, null);
+        assertTrue(isRowExpandedFired[0]);
+        // Row Long Press
+        final boolean[] isRowLongPressFired = {false};
+        table.addRowLongPressHandler((e, param1, param2, param3) -> {
+            isRowLongPressFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_LONGPRESS, null);
+        assertTrue(isRowLongPressFired[0]);
+        // Row Select
+        final boolean[] isRowSelectFired = {false};
+        table.addRowSelectHandler((e, param1, param2, param3) -> {
+            isRowSelectFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_SELECT, null);
+        assertTrue(isRowSelectFired[0]);
+        // Row Short Press
+        final boolean[] isRowShortPressFired = {false};
+        table.addRowShortPressHandler((e, param1, param2, param3) -> {
+            isRowShortPressFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.ROW_SHORTPRESS, null);
+        assertTrue(isRowShortPressFired[0]);
+        // Select All
+        final boolean[] isSelectAllFired = {false};
+        table.addSelectAllHandler((e, param1, param2, param3) -> {
+            isSelectAllFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.SELECT_ALL, null);
+        assertTrue(isSelectAllFired[0]);
+        // Sort Column
+        final boolean[] isSortColumnFired = {false};
+        table.addSortColumnHandler((e, param1, param2) -> {
+            isSortColumnFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.SORT_COLUMN, null);
+        assertTrue(isSortColumnFired[0]);
+        // Stretch
+        final boolean[] isStretchFired = {false};
+        table.addStretchHandler((e, param1) -> {
+            isStretchFired[0] = true;
+            return true;
+        });
+        $(table).trigger(TableEvents.STRETCH, null);
+        assertTrue(isStretchFired[0]);
+    }
+
+    public <T extends MaterialDataTable<Person>> void checkColumnsRows(T table) {
+        WidgetColumn<Person, MaterialImage> col1 = new WidgetColumn<Person, MaterialImage>() {
+            @Override
+            public MaterialImage getValue(Person object) {
+                MaterialImage profile = new MaterialImage();
+                profile.setUrl(object.getPicture());
+                profile.setWidth("40px");
+                profile.setHeight("40px");
+                profile.setPadding(4);
+                profile.setMarginTop(8);
+                profile.setBackgroundColor(Color.GREY_LIGHTEN_2);
+                profile.setCircle(true);
+                return profile;
+            }
+        };
+        table.addColumn(col1);
+        assertEquals(table.getColumns().get(0), col1);
+
+        Column<Person, ?> col2 = new TextColumn<Person>() {
+            @Override
+            public Comparator<? super RowComponent<Person>> sortComparator() {
+                return (o1, o2) -> o1.getData().getFirstName().compareToIgnoreCase(o2.getData().getFirstName());
+            }
+
+            @Override
+            public String getValue(Person object) {
+                return object.getFirstName();
+            }
+        };
+        table.addColumn(col2, "First Name");
+        assertEquals(table.getColumns().get(1), col2);
+        assertEquals(col2.getName(), "First Name");
+        assertTrue(col2.isSortable());
+
+        Column<Person, ?> col3 = new TextColumn<Person>() {
+            @Override
+            public Comparator<? super RowComponent<Person>> sortComparator() {
+                return (o1, o2) -> o1.getData().getLastName().compareToIgnoreCase(o2.getData().getLastName());
+            }
+
+            @Override
+            public String getValue(Person object) {
+                return object.getLastName();
+            }
+        };
+        table.addColumn(col3, "Last Name");
+        assertEquals(table.getColumns().get(2), col3);
+        assertEquals(col3.getName(), "Last Name");
+        assertTrue(col2.isSortable());
+
+        Column<Person, ?> col4 = new TextColumn<Person>() {
+            @Override
+            public boolean numeric() {
+                return true;
+            }
+
+            @Override
+            public HideOn hideOn() {
+                return HideOn.HIDE_ON_MED_DOWN;
+            }
+
+            @Override
+            public Comparator<? super RowComponent<Person>> sortComparator() {
+                return (o1, o2) -> o1.getData().getPhone().compareToIgnoreCase(o2.getData().getPhone());
+            }
+
+            @Override
+            public String getValue(Person object) {
+                return object.getPhone();
+            }
+        };
+        table.addColumn(col4, "Phone");
+        assertEquals(table.getColumns().get(3), col4);
+        assertEquals(col4.getName(), "Phone");
+        assertTrue(col4.isNumeric());
+
+        for (int i = 0; i < 4; i++) {
+            final int index = i;
+            Column<Person, ?> col = new TextColumn<Person>() {
+                @Override
+                public Comparator<? super RowComponent<Person>> sortComparator() {
+                    return (o1, o2) -> o1.getData().getPhone().compareToIgnoreCase(o2.getData().getPhone());
+                }
+
+                @Override
+                public String getValue(Person object) {
+                    return object.getPhone() + " " + index;
+                }
+            };
+            table.addColumn(col, "Column " + index);
+            assertEquals(table.getColumns().get(3 + (i + 1)), col);
+            assertEquals(col.getName(), "Column " + index);
+            assertTrue(col.isSortable());
+        }
+
+        Column<Person, ?> lastCol = new WidgetColumn<Person, MaterialBadge>() {
+            @Override
+            public TextAlign textAlign() {
+                return TextAlign.CENTER;
+            }
+
+            @Override
+            public MaterialBadge getValue(Person object) {
+                MaterialBadge badge = new MaterialBadge();
+                badge.setText("badge " + object.getId());
+                badge.setBackgroundColor(Color.BLUE);
+                badge.setLayoutPosition(Style.Position.RELATIVE);
+                return badge;
+            }
+        };
+        table.addColumn(lastCol);
+        assertEquals(table.getColumns().get(8), lastCol);
+
+        assertEquals(table.getColumns().size(), 9);
+
+        table.setVisibleRange(0, 1000);
+        Range range = table.getVisibleRange();
+        assertEquals(range.getStart(), 0);
+        assertEquals(range.getLength(), 1000);
+
+        List<Person> people = getAllPerson();
+        table.setRowData(0, people);
     }
 
     public <T extends MaterialDataTable<Person>> void checkStructure(T table) {
