--- src/test/java/gwt/material/design/client/test/MaterialDataTableTest.java
@@ -1,366 +1,0 @@
-package gwt.material.design.client.test;
-
-import com.google.gwt.dom.client.Style;
-import com.google.gwt.user.client.ui.RootPanel;
-import com.google.gwt.user.client.ui.Widget;
-import com.google.gwt.view.client.Range;
-import gwt.material.design.client.MaterialDataTableTestCase;
-import gwt.material.design.client.base.constants.TableCssName;
-import gwt.material.design.client.constants.Color;
-import gwt.material.design.client.constants.HideOn;
-import gwt.material.design.client.constants.IconType;
-import gwt.material.design.client.constants.TextAlign;
-import gwt.material.design.client.data.SelectionType;
-import gwt.material.design.client.data.component.RowComponent;
-import gwt.material.design.client.model.Person;
-import gwt.material.design.client.ui.MaterialBadge;
-import gwt.material.design.client.ui.MaterialCheckBox;
-import gwt.material.design.client.ui.MaterialDropDown;
-import gwt.material.design.client.ui.MaterialIcon;
-import gwt.material.design.client.ui.MaterialImage;
-import gwt.material.design.client.ui.table.MaterialDataTable;
-import gwt.material.design.client.ui.table.TableEvents;
-import gwt.material.design.client.ui.table.TableScaffolding;
-import gwt.material.design.client.ui.table.cell.Column;
-import gwt.material.design.client.ui.table.cell.TextColumn;
-import gwt.material.design.client.ui.table.cell.WidgetColumn;
-
-import java.util.Comparator;
-
-import static gwt.material.design.jquery.client.api.JQuery.$;
-
-public class MaterialDataTableTest extends MaterialDataTableTestCase {
-
-    public void testLoadingInConstructor() throws Exception {
-        // given / when / then
-        MaterialDataTable<Person> table = attachTableWithConstructor();
-
-        // Test the reattach
-        table.removeFromParent();
-        RootPanel.get().add(table);
-    }
-
-    public void testLoadingOnAttach() throws Exception {
-        // given / when / then
-        MaterialDataTable<Person> table = attachTableWithOnLoad();
-
-        // Test the reattach
-        table.removeFromParent();
-        RootPanel.get().add(table);
-    }
-
-    public void testProperties() throws Exception {
-        // given
-        MaterialDataTable<Person> table = attachTableWithOnLoad();
-
-        // when / then
-        table.setUseStickyHeader(true);
-        assertTrue(table.isUseStickyHeader());
-        table.setUseStickyHeader(false);
-        assertFalse(table.isUseStickyHeader());
-        table.setUseCategories(true);
-        assertTrue(table.isUseCategories());
-        table.setUseCategories(false);
-        assertFalse(table.isUseCategories());
-        table.setUseRowExpansion(true);
-        assertTrue(table.isUseRowExpansion());
-        table.setUseRowExpansion(false);
-        assertFalse(table.isUseRowExpansion());
-        table.setSelectionType(SelectionType.SINGLE);
-        assertEquals(table.getSelectionType(), SelectionType.SINGLE);
-        table.setSelectionType(SelectionType.MULTIPLE);
-        assertEquals(table.getSelectionType(), SelectionType.MULTIPLE);
-        table.setSelectionType(SelectionType.NONE);
-        assertEquals(table.getSelectionType(), SelectionType.NONE);
-    }
-
-    public void testStructure() throws Exception {
-        // given
-        MaterialDataTable<Person> table = attachTableWithOnLoad();
-
-        final String TABLE_TITLE = "table title";
-        // Table Title
-        table.setTitle(TABLE_TITLE);
-        assertEquals(table.getTitle(), TABLE_TITLE);
-        TableScaffolding scaffolding = table.getScaffolding();
-        MaterialIcon tableIcon = table.getTableIcon();
-        assertEquals(tableIcon.getIconType(), IconType.VIEW_LIST);
-        assertTrue(scaffolding.getInfoPanel().getElement().hasClassName(TableCssName.INFO_PANEL));
-        assertTrue(scaffolding.getTable().getElement().hasClassName(TableCssName.TABLE));
-        assertTrue(scaffolding.getTableBody().getElement().hasClassName(TableCssName.TABLE_BODY));
-        assertTrue(scaffolding.getToolPanel().getElement().hasClassName(TableCssName.TOOL_PANEL));
-        // Stretch Icon
-        MaterialIcon stretchIcon = table.getStretchIcon();
-        assertEquals(stretchIcon.getIconType(), IconType.FULLSCREEN);
-        assertEquals(stretchIcon.getId(), "stretch");
-        // Column Menu Icon
-        MaterialIcon columnMenuIcon = table.getColumnMenuIcon();
-        assertEquals(columnMenuIcon.getIconType(), IconType.MORE_VERT);
-        assertEquals(columnMenuIcon.getId(), "columnToggle");
-        assertTrue(scaffolding.getTopPanel().getElement().hasClassName(TableCssName.TOP_PANEL));
-        // Dropdown Menu
-        MaterialDropDown dropDown = table.getMenu();
-        assertEquals(dropDown.getWidgetCount(), 9);
-        int index = 0;
-        for (Widget w : dropDown) {
-            assertTrue(w instanceof MaterialCheckBox);
-            MaterialCheckBox checkBox = (MaterialCheckBox) w;
-            assertEquals(table.getColumns().get(index).getName(), checkBox.getText());
-            index++;
-        }
-    }
-
-    public void testEventHandlers() throws Exception {
-        // given
-        MaterialDataTable<Person> table = attachTableWithOnLoad();
-
-        // when / then
-
-        // Category Closed
-        final boolean[] isCategoryClosedFired = {false};
-        table.addCategoryClosedHandler((e, param1) -> {
-            isCategoryClosedFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.CATEGORY_CLOSED, null);
-        assertTrue(isCategoryClosedFired[0]);
-        // Category Opened
-        final boolean[] isCategoryOpenedFired = {false};
-        table.addCategoryOpenedHandler((e, param1) -> {
-            isCategoryOpenedFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.CATEGORY_OPENED, null);
-        assertTrue(isCategoryOpenedFired[0]);
-        // Row Context Menu
-        final boolean[] isRowContextMenuFired = {false};
-        table.addRowContextMenuHandler((e, param1, param2, param3) -> {
-            isRowContextMenuFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_CONTEXTMENU, null);
-        assertTrue(isRowContextMenuFired[0]);
-        // Row Count Change
-        final boolean[] isRowCountChangeFired = {false};
-        table.addRowCountChangeHandler((e, param1, param2) -> {
-            isRowCountChangeFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_COUNT_CHANGE, null);
-        assertTrue(isRowCountChangeFired[0]);
-        // Row Double Click
-        final boolean[] isRowDoubleClickFired = {false};
-        table.addRowDoubleClickHandler((e, param1, param2, param3) -> {
-            isRowDoubleClickFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_DOUBLECLICK, null);
-        assertTrue(isRowDoubleClickFired[0]);
-        // Row Expand
-        final boolean[] isRowExpandFired = {false};
-        table.addRowExpandHandler((e, param1) -> {
-            isRowExpandFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_EXPAND, null);
-        assertTrue(isRowExpandFired[0]);
-        // Row Expanded
-        final boolean[] isRowExpandedFired = {false};
-        table.addRowExpandedHandler((e, param1) -> {
-            isRowExpandedFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_EXPANDED, null);
-        assertTrue(isRowExpandedFired[0]);
-        // Row Long Press
-        final boolean[] isRowLongPressFired = {false};
-        table.addRowLongPressHandler((e, param1, param2, param3) -> {
-            isRowLongPressFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_LONGPRESS, null);
-        assertTrue(isRowLongPressFired[0]);
-        // Row Select
-        final boolean[] isRowSelectFired = {false};
-        table.addRowSelectHandler((e, param1, param2, param3) -> {
-            isRowSelectFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_SELECT, null);
-        assertTrue(isRowSelectFired[0]);
-        // Row Short Press
-        final boolean[] isRowShortPressFired = {false};
-        table.addRowShortPressHandler((e, param1, param2, param3) -> {
-            isRowShortPressFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.ROW_SHORTPRESS, null);
-        assertTrue(isRowShortPressFired[0]);
-        // Select All
-        final boolean[] isSelectAllFired = {false};
-        table.addSelectAllHandler((e, param1, param2, param3) -> {
-            isSelectAllFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.SELECT_ALL, null);
-        assertTrue(isSelectAllFired[0]);
-        // Sort Column
-        final boolean[] isSortColumnFired = {false};
-        table.addSortColumnHandler((e, param1, param2) -> {
-            isSortColumnFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.SORT_COLUMN, null);
-        assertTrue(isSortColumnFired[0]);
-        // Stretch
-        final boolean[] isStretchFired = {false};
-        table.addStretchHandler((e, param1) -> {
-            isStretchFired[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.STRETCH, null);
-        // Components Rendered
-        final boolean[] isCompRendered = {false};
-        table.addComponentsRenderedHandler(e -> {
-            isCompRendered[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.COMPONENTS_RENDERED, null);
-        assertTrue(isCompRendered[0]);
-        // Rendered
-        final boolean[] isRendered = {false};
-        table.addRenderedHandler(e -> {
-            isRendered[0] = true;
-            return true;
-        });
-        $(table).trigger(TableEvents.RENDERED, null);
-        assertTrue(isRendered[0]);
-    }
-
-    public <T extends MaterialDataTable<Person>> void testColumnsRows() throws Exception {
-        // given
-        MaterialDataTable<Person> table = attachTableWithOnLoad();
-        table.clearRows(true);
-        table.removeColumns();
-
-        WidgetColumn<Person, MaterialImage> col1 = new WidgetColumn<Person, MaterialImage>() {
-            @Override
-            public MaterialImage getValue(Person object) {
-                MaterialImage profile = new MaterialImage();
-                profile.setUrl(object.getPicture());
-                profile.setWidth("40px");
-                profile.setHeight("40px");
-                profile.setPadding(4);
-                profile.setMarginTop(8);
-                profile.setBackgroundColor(Color.GREY_LIGHTEN_2);
-                profile.setCircle(true);
-                return profile;
-            }
-        };
-        table.addColumn(col1);
-        assertEquals(table.getColumns().get(0), col1);
-
-        Column<Person, ?> col2 = new TextColumn<Person>() {
-            @Override
-            public Comparator<? super RowComponent<Person>> sortComparator() {
-                return (o1, o2) -> o1.getData().getFirstName().compareToIgnoreCase(o2.getData().getFirstName());
-            }
-
-            @Override
-            public String getValue(Person object) {
-                return object.getFirstName();
-            }
-        };
-        table.addColumn(col2, "First Name");
-        assertEquals(table.getColumns().get(1), col2);
-        assertEquals(col2.getName(), "First Name");
-        assertTrue(col2.isSortable());
-
-        Column<Person, ?> col3 = new TextColumn<Person>() {
-            @Override
-            public Comparator<? super RowComponent<Person>> sortComparator() {
-                return (o1, o2) -> o1.getData().getLastName().compareToIgnoreCase(o2.getData().getLastName());
-            }
-
-            @Override
-            public String getValue(Person object) {
-                return object.getLastName();
-            }
-        };
-        table.addColumn(col3, "Last Name");
-        assertEquals(table.getColumns().get(2), col3);
-        assertEquals(col3.getName(), "Last Name");
-        assertTrue(col2.isSortable());
-
-        Column<Person, ?> col4 = new TextColumn<Person>() {
-            @Override
-            public boolean numeric() {
-                return true;
-            }
-
-            @Override
-            public HideOn hideOn() {
-                return HideOn.HIDE_ON_MED_DOWN;
-            }
-
-            @Override
-            public Comparator<? super RowComponent<Person>> sortComparator() {
-                return (o1, o2) -> o1.getData().getPhone().compareToIgnoreCase(o2.getData().getPhone());
-            }
-
-            @Override
-            public String getValue(Person object) {
-                return object.getPhone();
-            }
-        };
-        table.addColumn(col4, "Phone");
-        assertEquals(table.getColumns().get(3), col4);
-        assertEquals(col4.getName(), "Phone");
-        assertTrue(col4.isNumeric());
-
-        for (int i = 0; i < 4; i++) {
-            final int index = i;
-            Column<Person, ?> col = new TextColumn<Person>() {
-                @Override
-                public Comparator<? super RowComponent<Person>> sortComparator() {
-                    return (o1, o2) -> o1.getData().getPhone().compareToIgnoreCase(o2.getData().getPhone());
-                }
-
-                @Override
-                public String getValue(Person object) {
-                    return object.getPhone() + " " + index;
-                }
-            };
-            table.addColumn(col, "Column " + index);
-            assertEquals(table.getColumns().get(3 + (i + 1)), col);
-            assertEquals(col.getName(), "Column " + index);
-            assertTrue(col.isSortable());
-        }
-
-        Column<Person, ?> lastCol = new WidgetColumn<Person, MaterialBadge>() {
-            @Override
-            public TextAlign textAlign() {
-                return TextAlign.CENTER;
-            }
-
-            @Override
-            public MaterialBadge getValue(Person object) {
-                MaterialBadge badge = new MaterialBadge();
-                badge.setText("badge " + object.getId());
-                badge.setBackgroundColor(Color.BLUE);
-                return badge;
-            }
-        };
-        table.addColumn(lastCol);
-        assertEquals(table.getColumns().get(8), lastCol);
-
-        assertEquals(table.getColumns().size(), 9);
-
-        table.setVisibleRange(0, 1000);
-        Range range = table.getVisibleRange();
-        assertEquals(range.getStart(), 0);
-        assertEquals(range.getLength(), 1000);
-
-        table.setRowData(0, people);
-    }
-}
